OPQBot_Botoy/requirements.txt
botoy
pydantic<2.0
tenacity
tinydb
tinyrecord
httpx[socks]
Pillow
py-cpuinfo
ujson
psutil
lxml
PicImageSearch
imgkit